+++ src/collective/accordionviews/tests/test_view_accordion_folders_view.py
@@ -4,7 +4,7 @@ import unittest
 from plone import api
 from plone.app.testing import TEST_USER_ID, setRoles
 from zope.component import getMultiAdapter
-from zope.component.interfaces import ComponentLookupError
+from zope.interface.interfaces import ComponentLookupError
 
 from collective.accordionviews.testing import (
     COLLECTIVE_ACCORDIONVIEWS_FUNCTIONAL_TESTING,
@@ -25,19 +25,15 @@ class ViewsIntegrationTest(unittest.TestCase):
     def test_accordion_folders_view_is_registered(self):
         view = getMultiAdapter(
             (self.portal["other-folder"], self.portal.REQUEST),
-            name="accordion-folders-view",
+            name="accordion_folders_view",
         )
-        self.assertTrue(view.__name__ == "accordion-folders-view")
-        # self.assertTrue(
-        #     'Sample View' in view(),
-        #     'Sample View is not found in accordion-folders-view'
-        # )
+        self.assertTrue(view.__name__ == "accordion_folders_view")
 
     def test_accordion_folders_view_not_matching_interface(self):
         with self.assertRaises(ComponentLookupError):
             getMultiAdapter(
                 (self.portal["front-page"], self.portal.REQUEST),
-                name="accordion-folders-view",
+                name="accordion_folders_view",
             )